% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\docType{data}
\name{birds}
\alias{birds}
\title{Example data}
\format{
dataframe with 30obs and 5 variables
\describe{
   \item{color}{color of bird}
  \item{country}{origin of bird}
  \item{diet}{diet of bird}
  \item{can_fly}{whether the bird can fly}
  \item{wingspan}{length of wingspan in cm}
  }
}
\usage{
birds
}
\description{
Birds data to be used for calculating cramersV scores
}
\keyword{datasets}
