% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cramers_v.R
\name{calculate_cramers_v}
\alias{calculate_cramers_v}
\title{Calculate Cramers V}
\usage{
calculate_cramers_v(df)
}
\arguments{
\item{df}{dataset}
}
\value{
CramersV score
}
\description{
Calculates CramersV score for categorical variables
}
\author{
Lily Zhou
}
