% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramerV-cohenW.R
\name{cohenW_df}
\alias{cohenW_df}
\title{Cohen's w for data.frames}
\usage{
cohenW_df(df, unique = TRUE, ci = FALSE, ...)
}
\arguments{
\item{df}{A data.table, tibble or data.frame with factor or character columns.}

\item{unique}{As the metric is symmetric (\code{cohenW(x,y) == cohenW(y,x)}) it avoids calculating twice the same value.}

\item{ci}{If TRUE, returns confidence intervals by bootstrap. May be slow.}

\item{...}{Additional arguments passed to \code{\link[rcompanion]{cohenW}}.}
}
\value{
A data.table, tibble or data.frame with the columns \code{V1}, \code{V2} and \code{Cohen.w}
with a row for each combination of categorical variables present
in the original \code{df}.

If \code{ci = TRUE} then \code{lower.ci} and \code{upper.ci} will be adding
based on the default \code{R} replications.
}
\description{
Calculates Cohen's w for each pair of categorical columns
in a data.frame, estimating confidence intervals by bootstrap.
}
\examples{
set.seed(2024)
df <- data.frame(
  gender = sample(c("male", "female"), 100, replace = TRUE) |> factor(),
  race = sample(c("white", "black", "asian", "other"), 100, replace = TRUE),
  education = sample(c("high school", "college", "graduate"), 100, replace = TRUE),
  income = sample(c("low", "medium", "high"), 100, replace = TRUE)
)
set.seed(NULL)

cohenW_df(df)
cohenW_df(df, ci = TRUE, R = 50)
}
\seealso{
\code{\link[rcompanion]{cohenW}} \code{\link[corrcat]{cramerV_df}}
}
\author{
Angel Feliz
}
