% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrcat-package.R
\docType{package}
\name{corrcat-package}
\alias{corrcat}
\alias{corrcat-package}
\title{corrcat: Tidy interface to 'rcompanion'}
\description{
corrcat facilitates to explore correlations between categorical and ordinal variables.
}
\author{
\strong{Maintainer}: Angel Feliz \email{angel.esteban.feliz@gmail.com}

}
\keyword{internal}
