% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmatrix_cat.R
\name{corrmatrix_cat}
\alias{corrmatrix_cat}
\title{Correlation Matrix for Categorical Variables}
\usage{
corrmatrix_cat(df)
}
\arguments{
\item{df}{dataset}
}
\value{
CramersV score and Correlation matrix
}
\description{
Calculates CramersV score for categorical variables and produces correlation matrix
}
\examples{
calculate_cramers_v(birds)

corrplot_cat(birds)
corrplot_cat(birds[1:4])
}
\author{
Lily Zhou
}
