% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot_cat.R
\name{corrplot_cat}
\alias{corrplot_cat}
\title{Correlation Matrix for Categorical Variables}
\usage{
corrplot_cat(df)
}
\arguments{
\item{df}{dataset}
}
\value{
CramersV score and Correlation matrix
}
\description{
Calculates CramersV score for categorical variables and produces correlation matrix.
Dataset should only include nominal categorical variables as Cramer's V does not work for continuous variables.
If continuous variables are included, correlation values produced will be falsely high.
}
\examples{

birdplot <- corrplot_cat(birds)
birdplot2 <- corrplot_cat(birds[1:4])
}
\author{
Lily Zhou
}
