% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_corrplot.R
\name{cramer_corrplot}
\alias{cramer_corrplot}
\title{Categorical Correlation Matrix}
\usage{
cramer_corrplot(df)
}
\arguments{
\item{df}{dataset}
}
\value{
Correlation Matrix
}
\description{
Generates correlation matrix from cramersV matrix
}
\author{
Lily Zhou
}
