% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot_cat.R
\name{corrplot_cat}
\alias{corrplot_cat}
\title{Categorical Correlation Matrix}
\usage{
corrplot_cat(df)
}
\arguments{
\item{df}{dataset}
}
\value{
Correlation Matrix
}
\description{
Generates correlation matrix from cramersV matrix
}
\author{
Lily Zhou
}
